[Gold IV] Title: 타일 채우기, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Gold/2133. 타일 채우기/타일 채우기.cc
@@ -8,7 +8,7 @@ int main()
     cin.tie(0);
     cout.tie(0);
 
-    long long dp[31] = {0};  // long long으로 변경하여 큰 수 처리
+    int dp[31] = {0};  // long long으로 변경하여 큰 수 처리
     int n;
     cin >> n;
